fix bug in timestamp calculation

+++ adma_ros_driver/src/adma_driver.cpp
@@ -186,8 +186,8 @@ void ADMADriver::parseData(std::array<char, 856> recv_buf)
         message_imu.header.seq = _seq;
         float weektime;
         //offset between UNIX and GNSS (in ms)
-        unsigned long offset_gps_unix = 315964800000;
-        uint32_t week_to_msec = 604800000;
+        unsigned long long offset_gps_unix = 315964800000;
+        unsigned long long week_to_msec = 604800000;
         unsigned long long timestamp;
 
         if (_protocol_version == "v3.3.3")
